examples/requirements.txt
browser-use[memory]==0.2.5
playwright>=1.46
python-dotenv>=1.0
